Add benchmarks for chalk 5 and yoctocolors
The latest version of `chalk` does not support CJS, but it is possible
to bundle it in the CJS format. It is worth testing the latest available
version.

`yoctocolors` does not support CJS, but it is possible to bundle it
in the CJS format. This library aspires to be the new tiniest one.
It supports `FORCE_COLOR` and `NO_COLOR` too. It is worth comparing
it to `picocolors`.

Split `chalk` to `chalk4` and `chalk5` using NPM module aliases.
Also upgrade development dependencies. Keep the same major version
of `clean-publish` not to break the legacy Node.js tests.

Use `esbuild` to create CJS bundles of `chalk5` and `yoctocolors`.
Add `chalk5` and `yoctocolors` to every benchmark. Rename the former
`chalk` tests to `chalk4`.

Correct the number formatting to allow one more decimal place.